Modifico en C# int por float en monto total factura

+++ Modelo/Alta.cs
@@ -41,7 +41,7 @@ namespace PagoAgilFrba.Modelo
             form.borrarItemsEnCreacion();
         }
 
-        public int calcularMontoTotalFactura(AltaModifFactura form)
+        public double calcularMontoTotalFactura(AltaModifFactura form)
         {
             return form.calcularMontoTotalEnCreacion();
         }
